Quick clean up on the configs.
