--- networktables.json.bck
@@ -1,3 +1,0 @@
-[
-
-]
